requirements.txt
fastapi[all]
paramiko
sqlalchemy
psycopg2
python-dotenv
requests
bandit
pysftp
bcrypt==3.2.0
passlib[bcrypt]
pylint
pyjwt[crypto]
pytest
httpx
apscheduler
configparser
ldap3
ping3
pyOpenSSL==24.3.0
pillow